fix register function

+++ backend/service/user-service.ts
@@ -6,6 +6,8 @@ import { PasswordEncoder } from "./password-encoder";
 import { Injectable, DIProperty } from "../lib/di";
 import { v4 as uuidv4 } from "uuid";
 import { JWTService } from "./jwt-service";
+import createHttpError from "http-errors";
+import path from "path";
 
 
 interface UserBasicInfo {
@@ -59,16 +61,71 @@ export class UserService {
         const code = this.create_uuid();
         const inv = new InvitationCode();
         inv.code = code;
-        inv.user = user;
+        inv.user = Promise.resolve(user);
+        inv.userId = user.id;
         inv.relativepath = ".";
         await this.invRepo.save(inv);
         return code;
     }
 
-    public async deleteInvitationCode(code: string) {
-        return await this.invRepo.delete({
-            code: code
+    public async deleteInvitationCode(username: string, code: string): Promise<void> {
+        const user = await this.getUser(username);
+        if (!user)
+            throw new createHttpError.InternalServerError();
+        const m = await this.invRepo.delete({
+            code: code,
+            userId: user.id,
         });
+        if (m.affected == 0)
+            throw new createHttpError.NotFound();
+    }
+
+    public async getInvitationCode(username: string): Promise<{ code: string, relpath: string } []> {
+        const user = await this.getUser(username);
+        if (!user) {
+            throw new createHttpError.InternalServerError("User not found");
+        }
+        const codes = await user.invitationCodes;
+        return codes.map(c => { return {code: c.code, relpath: c.relativepath}; });
+    }
+
+    public async getInvCode(username: string, code: string): Promise<InvitationCode> {
+        const user = await this.getUser(username);
+        if (!user) {
+            throw new createHttpError.InternalServerError("User not found");
+        }
+        const val =  await this.invRepo
+            .createQueryBuilder("invcode")
+            .select()
+            .where("invcode.userId = :userId", {userId: user.id})
+            .andWhere("invcode.code = :ic", {ic: code})
+            .getOne();
+        if (!val) {
+            throw new createHttpError.NotFound("invitation code not found");
+        }
+        return val;
+    }
+
+    public async saveInvCode(invcode: InvitationCode): Promise<void> {
+        await this.invRepo.save(invcode);
+    }
+
+    public async getUserInfoByInvCode(_username: string, code: string): Promise<{ [key: string]: string }> {
+        const user = await this.userRepo
+            .createQueryBuilder("user")
+            .select()
+            .innerJoin(InvitationCode, "inv", "user.selfInvitationCodeId = inv.id")
+            .where("inv.code = :code", { code: code })
+            .getOne();
+
+        if (!user) {
+            return {};
+        } else {
+            return { 
+                createdAt: user.createdAt.toUTCString(),
+                username: user.username,
+            };
+        }
     }
 
     public async getProfilePicture(username: string): Promise<string> {
@@ -135,24 +192,32 @@ export class UserService {
     }
 
     public async createUser(username: string, password: string, inv: string) {
+        if (!username || !password || !inv)
+            throw new createHttpError.InternalServerError();
+
         await this.dataSource.transaction(async (manager) => {
             const up = manager.getRepository(User);
             const ip = manager.getRepository(InvitationCode);
             const sp = manager.getRepository(UserSetting);
 
-            let invCode = null;
-            if (inv != null) {
-                invCode = await ip.findOne({
-                    where: {
-                        code: inv
-                    }
-                });
-                if (!invCode) {
-                    throw new createError.NotFound("Invalid invitation code");
+            const invCode = await ip.findOne({
+                where: {
+                    code: inv
                 }
+            });
+            if (!invCode) {
+                throw new createError.NotFound("Invalid invitation code");
             }
 
-            const oldUser = up.findOne({
+            const existUser = await up.createQueryBuilder("user").select().where({
+                selfInvitationCode: invCode,
+            }).getOne();
+            if (existUser)
+                throw new createHttpError.Conflict("invitation code already be used");
+
+            const inviterUser = await invCode.user;
+
+            const oldUser = await up.findOne({
                 where: {
                     username: username
                 }
@@ -164,14 +229,16 @@ export class UserService {
             const user = new User();
             user.username = username;
             user.password = this.encoder.encode(password);
-            user.selfInvitationCode = invCode;
-            up.save(user);
+            user.selfInvitationCode = Promise.resolve(invCode);
+            user.selfInvitationCodeId = invCode.id;
+            user.rootpath = path.join(inviterUser.rootpath, invCode.relativepath);
+            await up.save(user);
 
             const setting = new UserSetting();
             setting.user = user;
             setting.frontendSettings = "{}";
             setting.profilePicture = "";
-            sp.save(setting);
+            await sp.save(setting);
         });
     }
 
@@ -216,7 +283,7 @@ export class UserService {
             .createQueryBuilder()
             .select("user.id")
             .from(User, "user")
-            .innerJoin(InvitationCode, "inv", "inv.user = user.id")
+            .innerJoin(InvitationCode, "inv", "inv.id = user.selfInvitationCodeId")
             .where("user.username = :username", {username: username})
             .andWhere("inv.code = :invcode", {invcode: invcode})
             .getOne() != null;
@@ -243,10 +310,11 @@ export class UserService {
         const token = this.create_uuid();
         const refreshToken = new RefreshToken();
         refreshToken.token = token;
-        refreshToken.user = user;
+        refreshToken.user = Promise.resolve(user);
+        refreshToken.userId = user.id;
         refreshToken.updateCount = 0;
         await this.refRepo.save(refreshToken);
-        await this.refRepo.update(refreshToken, {updateCount: refreshToken.updateCount+1});
+        await this.refRepo.update({token: token}, {updateCount: refreshToken.updateCount+1});
         return token;
     }
 
@@ -290,8 +358,9 @@ export class UserService {
             throw createError(401, "Token expired");
         }
         await this.refRepo.update({ id: refreshToken.id }, { updateCount: refreshToken.updateCount+1 });
-        const username = refreshToken.user.username;
-        const rootpath = refreshToken.user.rootpath;
+        const user = await refreshToken.user;
+        const username = user.username;
+        const rootpath = user.rootpath;
         return this.jwtService.sign({username: username, rootpath: rootpath}, username);
     }
 }
